changes to ease use of WALA from an executable jar.  also some very basic support for GXL, a graph format used by some software engineering tools.
