+++ com.ibm.wala.cast.js.rhino/META-INF/MANIFEST.MF
@@ -35,5 +35,5 @@ Export-Package: com.ibm.wala.cast.js.translator,
  org.mozilla.javascript.xml,
  org.mozilla.javascript.xml.impl.xmlbeans,
  org.mozilla.javascript.xmlimpl
-Bundle-ClassPath: lib/rhino-1.7R3.jar,
- .
+Bundle-ClassPath: .,
+ lib/js.jar
