+++ com.ibm.wala.cast.js.rhino/build.properties
@@ -2,5 +2,5 @@ source.. = source/
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
-               lib/rhino-1.7R3.jar
+               lib/js.jar
 jars.extra.classpath = lib/rhino-1.7R3.jar
